requirements.txt
inflection
requests
python-dotenv